+++ src/RFSpark/RFSpark_wxgui.cpp
@@ -26,7 +26,7 @@ const long RFSpark_wxgui::ID_BTNREADGPIO = wxNewId();
 const long RFSpark_wxgui::ID_CMBSELECTADC = wxNewId();
 
 BEGIN_EVENT_TABLE(RFSpark_wxgui,wxFrame)
-	
+
 END_EVENT_TABLE()
 
 wxString power2unitsString(char powerx3)
@@ -79,7 +79,7 @@ RFSpark_wxgui::RFSpark_wxgui(wxWindow* parent,wxWindowID id, const wxString& tit
     SetIcon(wxIcon(_("aaaaAPPicon")));
 #endif
 
-    //ADC values	
+    //ADC values
     wxFlexGridSizer* sizerADCs = new wxFlexGridSizer(0, 3, 0, 10);
 
     ADCdataGUI adcElement;
@@ -202,7 +202,7 @@ void RFSpark_wxgui::UpdateADClabels()
 	{
             mADCgui[i].value->SetLabelText(wxString::Format("%f", mADCdata[i].value));
             mADCgui[i].units->SetLabelText(wxString::Format("%s", mADCdata[i].units));
-	
+
             /*
 		mADCgui[i].value->SetLabelText(wxString::Format("%i", mADCdata[i].value));
 		mADCgui[i].units->SetLabelText(wxString::Format("%s%s", power2unitsString(mADCdata[i].powerOf10coefficient), adcUnits2string(mADCdata[i].units)));
@@ -212,8 +212,7 @@ void RFSpark_wxgui::UpdateADClabels()
 
 void RFSpark_wxgui::OnbtnRefreshAllADC(wxCommandEvent& event)
 {
-  
-    for (int i = 0; i < mADCdata.size(); ++i)
+    for (size_t i = 0; i < mADCdata.size(); ++i)
     {
         float_type val;
         lms_name_t units;
@@ -226,12 +225,12 @@ void RFSpark_wxgui::OnbtnRefreshAllADC(wxCommandEvent& event)
         mADCdata[i].units = units;
         mADCdata[i].value = val;
     }
-    
+
     UpdateADClabels();
-   /* 
+   /*
     LMS64CProtocol::GenericPacket pkt;
     pkt.cmd = CMD_ANALOG_VAL_RD;
-    
+
     for (int i = 0; i < mADCdata.size(); ++i)
         pkt.outBuffer.push_back(i);
 
@@ -240,7 +239,7 @@ void RFSpark_wxgui::OnbtnRefreshAllADC(wxCommandEvent& event)
         wxMessageBox(_("Board response: ") + wxString::From8BitData(GetLastErrorMessage()), _("Warning"));
         return;
     }
-	
+
     assert(pkt.inBuffer.size() >= 4 * mADCcount);
 
 	for (int i = 0; i < mADCdata.size(); ++i)
@@ -248,7 +247,7 @@ void RFSpark_wxgui::OnbtnRefreshAllADC(wxCommandEvent& event)
 		mADCdata[i].channel = pkt.inBuffer[i*4];
         mADCdata[i].units = (pkt.inBuffer[i * 4 + 1] & 0xF0) >> 4;
         mADCdata[i].powerOf10coefficient = pkt.inBuffer[i * 4 + 1] << 4;
-		mADCdata[i].powerOf10coefficient = mADCdata[i].powerOf10coefficient >> 4;				
+		mADCdata[i].powerOf10coefficient = mADCdata[i].powerOf10coefficient >> 4;
         mADCdata[i].value = pkt.inBuffer[i * 4 + 2] << 8 | pkt.inBuffer[i * 4 + 3];
 	}
 	UpdateADClabels();
@@ -258,7 +257,7 @@ void RFSpark_wxgui::OnbtnRefreshAllADC(wxCommandEvent& event)
 void RFSpark_wxgui::OnbtnRefreshADC(wxCommandEvent& event)
 {
 
-    int index = cmbADCselect->GetSelection(); 
+    int index = cmbADCselect->GetSelection();
 
     float_type val;
     lms_name_t units;
@@ -270,14 +269,14 @@ void RFSpark_wxgui::OnbtnRefreshADC(wxCommandEvent& event)
     mADCdata[index].channel = index;
     mADCdata[index].units = units;
     mADCdata[index].value = val;
-    
+
     UpdateADClabels();
     /*
     LMS64CProtocol::GenericPacket pkt;
     pkt.cmd = CMD_ANALOG_VAL_RD;
 
     int index = cmbADCselect->GetSelection();
-    
+
     pkt.outBuffer.push_back(index);
     assert(m_serPort != nullptr);
     if (m_serPort->TransferPacket(pkt) != 0)
@@ -285,7 +284,7 @@ void RFSpark_wxgui::OnbtnRefreshADC(wxCommandEvent& event)
         wxMessageBox(_("Board response: ") + wxString::From8BitData(GetLastErrorMessage()), _("Warning"));
         return;
     }
-	
+
     assert(pkt.inBuffer.size() >= 4);
 
 	if(index < mADCdata.size())
@@ -300,19 +299,19 @@ void RFSpark_wxgui::OnbtnRefreshADC(wxCommandEvent& event)
 }
 
 void RFSpark_wxgui::OnbtnWriteGPIO(wxCommandEvent& event)
-{  
-    
+{
+
 	uint8_t* values = new uint8_t[mGPIOboxes.size() / 8];
     int gpioIndex = 0;
-    for (int i = 0; i < mGPIOboxes.size()/8; ++i)
-    { 
+    for (size_t i = 0; i < mGPIOboxes.size()/8; ++i)
+    {
         unsigned char value = 0;
-        for (int j = 7; j >= 0; --j)		
-            value |= mGPIOboxes[gpioIndex++]->IsChecked() << j;			
+        for (int j = 7; j >= 0; --j)
+            value |= mGPIOboxes[gpioIndex++]->IsChecked() << j;
         values[i] = value;
-    }    
-    
-    if (LMS_GPIOWrite(lmsControl,values,mGPIOboxes.size()/8) != 0)		
+    }
+
+    if (LMS_GPIOWrite(lmsControl,values,mGPIOboxes.size()/8) != 0)
         wxMessageBox(_("Board response: ") + wxString::From8BitData(LMS_GetLastErrorMessage()), _("Warning"));
 
 	delete [] values;
@@ -320,37 +319,37 @@ void RFSpark_wxgui::OnbtnWriteGPIO(wxCommandEvent& event)
 /*
     LMS64CProtocol::GenericPacket pkt;
     pkt.cmd = CMD_GPIO_WR;
-    
+
 	int gpioIndex = 0;
 	for (int i = 0; i < mGPIOboxes.size()/8; ++i)
-	{ 
+	{
         unsigned char value = 0;
-		for (int j = 7; j >= 0; --j)		
-			value |= mGPIOboxes[gpioIndex++]->IsChecked() << j;			
+		for (int j = 7; j >= 0; --j)
+			value |= mGPIOboxes[gpioIndex++]->IsChecked() << j;
         pkt.outBuffer.push_back(value);
-	}    
-    if (m_serPort->TransferPacket(pkt) != 0)		
+	}
+    if (m_serPort->TransferPacket(pkt) != 0)
         wxMessageBox(_("Board response: ") + wxString::From8BitData(GetLastErrorMessage()), _("Warning"));
   */
 }
 
 void RFSpark_wxgui::OnbtnReadGPIO(wxCommandEvent& event)
 {
-    
+
     uint8_t* values = new uint8_t[mGPIOboxes.size()/8];
-    
-    if (LMS_GPIORead(lmsControl,values,mGPIOboxes.size()/8) != 0)	
+
+    if (LMS_GPIORead(lmsControl,values,mGPIOboxes.size()/8) != 0)
         wxMessageBox(_("Board response: ") + wxString::From8BitData(LMS_GetLastErrorMessage()), _("Warning"));
-    
-    int gpioIndex = 0;
+
+    size_t gpioIndex = 0;
     int gpioByte = 0;
-    for (int i = 0; i < mGPIOboxes.size() / 8; ++i)
-    {		
+    for (size_t i = 0; i < mGPIOboxes.size() / 8; ++i)
+    {
             for (int j = 7; j >= 0 && gpioIndex < mGPIOboxes.size(); --j)
                     mGPIOboxes[gpioIndex++]->SetValue( values[gpioByte] & (0x1 << j) );
             ++gpioByte;
     }
-    
+
 	delete[] values;
 /*
     LMS64CProtocol::GenericPacket pkt;
@@ -363,14 +362,14 @@ void RFSpark_wxgui::OnbtnReadGPIO(wxCommandEvent& event)
         return;
     }
     assert(pkt.inBuffer.size() >= 3);
-		
+
 	int gpioIndex = 0;
 	int gpioByte = 0;
 	for (int i = 0; i < mGPIOboxes.size() / 8; ++i)
-	{		
+	{
 		for (int j = 7; j >= 0 && gpioIndex < mGPIOboxes.size(); --j)
 			mGPIOboxes[gpioIndex++]->SetValue( pkt.inBuffer[gpioByte] & (0x1 << j) );
 		++gpioByte;
-	}	
+	}
  */
 }
